Move test and real startups to own files
